wifi: Use stub for add_or_remove_virtual_intf functionality
Use the stub to take care of cases where the vendor HAL
implementation is not present. This will return an error
as NOT_SUPPORTED when the functionality is needed by framework
(with no implementation wifi hal). This will avoid the failure
scenario when vendor implementation for add_or_remove_virtual_intf
is not present in wifi hal.

Change-Id: I5c1c99acb40e721e22afa2683b5b0836e533f34e
CRs-Fixed: 2363891